+++ sitzverteilung-app/apigateway/src/test/java/de/muenchen/sitzverteilungapp/TestConstants.java
@@ -2,7 +2,7 @@
  * Copyright (c): it@M - Dienstleister für Informations- und Telekommunikationstechnik
  * der Landeshauptstadt München, 2024
  */
-package de.muenchen.sitzverteilung-app;
+package de.muenchen.sitzverteilungapp;
 
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
