+++ sitzverteilung-app/apigateway/src/test/java/de/muenchen/sitzverteilungapp/route/BackendRouteTest.java
@@ -2,11 +2,12 @@
  * Copyright (c): it@M - Dienstleister für Informations- und Telekommunikationstechnik
  * der Landeshauptstadt München, 2024
  */
-package de.muenchen.sitzverteilung-app.filter;
+package de.muenchen.sitzverteilungapp.route;
 
-import de.muenchen.sitzverteilung-app.ApiGatewayApplication;
 import com.github.tomakehurst.wiremock.http.HttpHeader;
 import com.github.tomakehurst.wiremock.http.HttpHeaders;
+import com.github.tomakehurst.wiremock.matching.EqualToPattern;
+import de.muenchen.sitzverteilungapp.ApiGatewayApplication;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -19,11 +20,13 @@ import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 import org.springframework.test.web.reactive.server.WebTestClient;
 
-import static de.muenchen.sitzverteilung-app.TestConstants.SPRING_TEST_PROFILE;
+import static de.muenchen.sitzverteilungapp.TestConstants.SPRING_TEST_PROFILE;
 import static com.github.tomakehurst.wiremock.client.WireMock.aResponse;
 import static com.github.tomakehurst.wiremock.client.WireMock.get;
+import static com.github.tomakehurst.wiremock.client.WireMock.getRequestedFor;
 import static com.github.tomakehurst.wiremock.client.WireMock.stubFor;
 import static com.github.tomakehurst.wiremock.client.WireMock.urlEqualTo;
+import static com.github.tomakehurst.wiremock.client.WireMock.verify;
 
 @ExtendWith(SpringExtension.class)
 @SpringBootTest(
@@ -32,36 +35,42 @@ import static com.github.tomakehurst.wiremock.client.WireMock.urlEqualTo;
 )
 @ActiveProfiles(SPRING_TEST_PROFILE)
 @AutoConfigureWireMock
-class GlobalAuthenticationErrorFilterTest {
+class BackendRouteTest {
 
     @Autowired
     private WebTestClient webTestClient;
 
     @BeforeEach
     void setup() {
-        stubFor(get(urlEqualTo("/remote"))
+        stubFor(get(urlEqualTo("/remote/endpoint"))
                 .willReturn(aResponse()
-                        .withStatus(HttpStatus.UNAUTHORIZED.value())
+                        .withStatus(HttpStatus.OK.value())
                         .withHeaders(new HttpHeaders(
                                 new HttpHeader("Content-Type", "application/json"),
-                                new HttpHeader("WWW-Authenticate", "Bearer realm=\"Access to the staging site\", charset=\"UTF-8\""),
-                                new HttpHeader("Expires", "Wed, 21 Oct 2099 07:28:06 GMT")))
+                                new HttpHeader("WWW-Authenticate", "Bearer realm=\"Access to the staging site\", charset=\"UTF-8\""), // removed by route filter
+                                new HttpHeader("Expires", "Wed, 21 Oct 2099 07:28:06 GMT") // removed by route filter
+                        ))
                         .withBody("{ \"testkey\" : \"testvalue\" }")));
     }
 
     @Test
     @WithMockUser
-    void backendAuthenticationError() {
-        //@formatter:off
-        webTestClient.get().uri("/api/sitzverteilung-backend-service/remote").exchange()
-                .expectStatus().isEqualTo(HttpStatus.UNAUTHORIZED)
+    void backendRouteResponse() {
+        webTestClient.get().uri("/api/sitzverteilung-backend-service/remote/endpoint")
+                .header("Cookie", "SESSION=5cfb01a3-b691-4ca9-8735-a05690e6c2ec; XSRF-TOKEN=4d82f9f1-41f6-4a09-994a-df99d30d1be9") // removed by default-filter
+                .header("X-XSRF-TOKEN", "5cfb01a3-b691-4ca9-8735-a05690e6c2ec") // angular specific -> removed by default-filter
+                .header("Content-Type", "application/hal+json")
+                .exchange()
+                .expectStatus().isEqualTo(HttpStatus.OK)
                 .expectHeader().valueMatches("Content-Type", "application/json")
                 .expectHeader().doesNotExist("WWW-Authenticate")
                 .expectHeader().valueMatches("Expires", "0")
-                .expectBody()
-                    .jsonPath("$.status").isEqualTo("401")
-                    .jsonPath("$.error").isEqualTo("Authentication Error");
-        //@formatter:on
+                .expectBody().jsonPath("$.testkey").isEqualTo("testvalue");
+
+        verify(getRequestedFor(urlEqualTo("/remote/endpoint"))
+                .withoutHeader("Cookie")
+                .withoutHeader("X-SRF-TOKEN")
+                .withHeader("Content-Type", new EqualToPattern("application/hal+json")));
     }
 
 }
